refactor(server): move websocket and mutation policies into package
